Fixed using undeclared type, use ULONG instead
git-svn-id: https://svn.aros.org/svn/aros/trunk@38364 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/dos/internalloadseg_elf.c
@@ -507,7 +507,7 @@ static int relocate
                 if (offset & 0x02000000)
                     offset -= 0x04000000;
 
-                offset += s - (uint32_t)p;
+                offset += s - (ULONG)p;
 
                 offset >>= 2;
                 *p &= 0xff000000;
